Replace 'columns' generator with 'iter_rows'
The latest version of the library returns a generator with the `columns`
attribute. I replace this with a call to `iter_rows` to keep the exact same
semantic without going into why you would have to do `list(sheet.columns)[1]`.

An other alternative is to talk about generators and show how that would work
(potentially too advanced). Or we could rework the example to explain a bit
more about what `iter_rows` and `iter_cols` can do to slice your sheet.